+++ pyproject.toml
@@ -30,7 +30,6 @@ pinecone-client = "^2.2.1"
 tiktoken = "^0.4.0"
 pypdf = "^3.8.1"
 cohere = "^4.5.1"
-playwright = "^1.34.0"
 
 
 [build-system]
